feat: v0.8.0 — Massive mathematical engine upgrade
## Core Changes (35+ features)
- **CAS**: Symbolic differentiation (29 AST variants), Gauss-Legendre adaptive integration,
  Newton root finding, Richardson limits, Taylor series, matrix operations (det, inv, solve)
- **19 Probability Distributions**: Normal, Student-t, Chi-square, F, Exponential, Geometric,
  Hypergeometric, Logistic, Weibull, Uniform, Gamma, Beta, Cauchy, Pareto, Rayleigh, Laplace,
  NegBinomial — all with PDF/PMF + CDF, most with quantile
- **Statistical Inference**: t-test (1 & 2 sample), z-test, Chi-square test, ANOVA, confidence intervals
- **Regressions**: Linear (R²), polynomial (Gauss elim), exponential, logarithmic, power
- **Graphs**: Histogram, ScatterPlot, BoxPlot (outliers), RegressionLine, FrequencyTable
- **Descriptive Stats**: Mean, Median, Mode, StdDev, Variance, Quantiles, IQR, Covariance, Pearson

## Advanced Mathematics
- **ODE Solvers**: Euler + RK4 for single ODEs and 2D systems (phase portraits)
- **Special Functions**: Gamma, Beta, Bessel J/Y/I, Erf, Erfc, Digamma
- **Special Curves**: Cardioid, Rose, Archimedean/Logarithmic Spiral, Lissajous, Epicycloid, Hypocycloid
- **10 Strange Attractors**: Lorenz, Rössler, Thomas/Butterfly, Aizawa, Chen, Halvorsen, Dadras,
  Chua, Sprott, Three-Scroll — all with RK4 integration caching
- **Fractals**: Mandelbrot, Julia, Burning Ship, Tricorn, Newton — parallel via rayon
- **4D Projection**: Hypercube (tesseract), Hypersphere with rotation angles
- **3D Vector Fields**: Grid-based arrow visualization

## 3D Graphics
- **Depth sorting**: Painters algorithm for correct occlusion
- **Phong-style lighting**: Ambient + diffuse per-vertex on spheres, cubes, pyramids, cones, cylinders
- **8 solid types with normals**: Colored wireframes with face/surface lighting

## UI/UX
- **16-tool toolbar** with vector icons, contextual show/hide per 2D/3D mode
- **Command Palette (Ctrl+K)**: Fuzzy search across 70+ commands in 12 categories
- **Properties Panel**: Right-side panel with label editing, visibility, color, measurements
- **Algebra Panel**: Search bar, category filters (Todos, Puntos, Líneas, Curvas, Funciones, 3D)
- **Ghost preview**: Semi-transparent tool preview while hovering (Point, Line, Circle, Polygon)
- **Ripple effects**: Click animations on canvas
- **Math Keyboard**: 5-tab virtual keyboard including advanced commands tab
- **Dark/Light theme** toggle

## Tool Interaction Fixes
- Point tool: drag-release to place (eliminates click sensitivity bug)
- Drag only pans in Select mode (doesn't interfere with drawing tools)
- Polygon: right-click to close with 3+ vertices
- Right-click cancels single pending point for Line/Circle
- 3D tools: drag_stopped fallback for reliable click detection

## Performance
- **Attractor point caching**: Parameter hash → cached point cloud, ~100x speedup
- **Parallel fractals**: rayon::par_iter, 4-8x speedup on multi-core
- **Spatial index**: R-tree for 32 object types, O(log n) hit testing
- **2D/3D object filtering**: Objects hidden from iteration when not in current view mode

## Export
- SVG/PNG/TikZ export expanded: Ellipse, Parabola, Hyperbola, Function, Text, ScatterPlot, RegressionLine

## Bug Fixes
- AST 'e' preprocessing no longer corrupts identifiers
- log2 correctly maps to ln(x)/ln(2)
- student_t_cdf fixed for negative x
- geometric_pmf unified convention
- Dark mode labels now visible
- Document::clear() properly resets spatial index and spreadsheet
- 3D depth buffer incompatibility crash fixed
- Tests passing (74+ unit tests)

+++ crates/grafito-app/src/export.rs
@@ -1,3 +1,4 @@
+#![allow(dead_code)]
 use grafito_core::{Document, GeoObject};
 use grafito_geometry::{Point2, Color};
 use grafito_geometry::expr::eval_function_with_vars;
@@ -30,7 +31,7 @@ pub fn export_tikz(document: &Document) -> String {
             }
             GeoObject::Circle(c) => {
                 let cen = view.world_to_screen(c.center);
-                let r = c.radius as f32 * view.scale;
+                let r = c.radius * view.scale;
                 out.push_str(&format!("  \\draw ({:.2},{:.2}) circle ({:.2});\n", cen.x, cen.y, r));
             }
             GeoObject::Polygon(poly) if poly.vertices.len() >= 3 => {
@@ -42,7 +43,7 @@ pub fn export_tikz(document: &Document) -> String {
             GeoObject::Ellipse(el) => {
                 let cen = view.world_to_screen(el.center);
                 out.push_str(&format!("  \\draw ({:.2},{:.2}) ellipse ({:.2} and {:.2});\n",
-                    cen.x, cen.y, el.rx as f32 * view.scale, el.ry as f32 * view.scale));
+                    cen.x, cen.y, el.rx * view.scale, el.ry * view.scale));
             }
             _ => {}
         }
@@ -108,13 +109,13 @@ pub fn export_svg(document: &Document) -> String {
             }
             GeoObject::Circle(c) => {
                 let center = view.world_to_screen(c.center);
-                let r = (c.radius as f32) * view.scale;
+                let r = c.radius * view.scale;
                 if let Some(fill) = c.fill_color {
                     let _ = writeln!(svg, r#"<circle cx="{:.1}" cy="{:.1}" r="{:.1}" fill="{}"/>"#, center.x, center.y, r, svg_color(fill));
                 }
                 let _ = writeln!(svg, r#"<circle cx="{:.1}" cy="{:.1}" r="{:.1}" fill="none" stroke="{}" stroke-width="{}"/>"#, center.x, center.y, r, svg_color(c.color), c.width);
                 if !c.label.is_empty() {
-                    let _ = writeln!(svg, r#"<text x="{:.1}" y="{:.1}" font-family="sans-serif" font-size="12" fill="black">{}</text>"#, center.x + r + 2.0, center.y - r - 2.0, c.label);
+                    let _ = writeln!(svg, r#"<text x="{:.1}" y="{:.1}" font-family="sans-serif" font-size="12" fill="black">{}</text>"#, center.x + r as f32 + 2.0, center.y - r as f32 - 2.0, c.label);
                 }
             }
             GeoObject::Polygon(poly) => {
@@ -233,7 +234,7 @@ pub fn export_png(document: &Document, width: u32, height: u32) -> image::RgbaIm
             }
             GeoObject::Circle(c) => {
                 let center = view.world_to_screen(c.center);
-                let r = (c.radius as f32 * view.scale).max(0.5) as i32;
+                let r = (c.radius * view.scale).max(0.5) as i32;
                 if let Some(fill) = c.fill_color {
                     let fc = [(fill.r*255.0) as u8, (fill.g*255.0) as u8, (fill.b*255.0) as u8, 255];
                     fill_circle(&mut img, center.x as i32, center.y as i32, r, fc);
